+++ src/IQFeed.CSharpApiClient.Tests.Integration/Lookup/Historical/HistoricalFileFacadeTests.cs
@@ -42,8 +42,7 @@ namespace IQFeed.CSharpApiClient.Tests.Integration.Lookup.Historical
         [Test, MaxTime(TimeoutMs)]
         public async Task Should_Return_String_When_ReqHistoryTickDaysAsync()
         {
-            // 6.1 protocol only allows Int16.MaxValue days to be requested
-            var tmpFilename = await _lookupClient.Historical.File.GetHistoryTickDaysAsync(Symbol, Int16.MaxValue, Datapoints);
+            var tmpFilename = await _lookupClient.Historical.File.GetHistoryTickDaysAsync(Symbol, int.MaxValue, Datapoints);
             Assert.IsNotEmpty(tmpFilename);
         }
 
@@ -64,8 +63,7 @@ namespace IQFeed.CSharpApiClient.Tests.Integration.Lookup.Historical
         [Test, MaxTime(TimeoutMs)]
         public async Task Should_Return_String_When_ReqHistoryIntervalDaysAsync()
         {
-            // 6.1 protocol only allows Int16.MaxValue days to be requested
-            var tmpFilename = await _lookupClient.Historical.File.GetHistoryIntervalDaysAsync(Symbol, 5, Int16.MaxValue, Datapoints);
+            var tmpFilename = await _lookupClient.Historical.File.GetHistoryIntervalDaysAsync(Symbol, 5, int.MaxValue, Datapoints);
             Assert.IsNotEmpty(tmpFilename);
         }
